Refactor table contents handling, remove obsolete files, and add new tests for MCP server functionality
