+++ .gitignore
@@ -12,3 +12,4 @@ __pycache__/
 *.egg
 *.zip
 *.env
+sk.json
